requirements.txt
fastapi
uvicorn[standard]
openai
pinecone-client
redis
python-dotenv
requests
pydantic
pytest
